Implement TokenBuilder and UserFinder
Created a POCO called UserCredentials that will be used later as a way
to reference incoming data from login and register requests.

Implemented UserFinder service that takes a UserCredentials object and
attempts to find an ApplicationUser with the given credentials.

Implemented TokenBuilder service that creates a Javascript Web Token for
an ApplicationUser.